+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeMirror.java
@@ -642,7 +642,7 @@ public abstract class AnnotatedTypeMirror implements DeepCopyable<AnnotatedTypeM
      * @see #getExplicitAnnotations()
      */
     public boolean hasExplicitAnnotation(Class<? extends Annotation> a) {
-        return AnnotationUtils.containsSameByName(getExplicitAnnotations(), getAnnotation(a));
+        return AnnotationUtils.containsSameByClass(getExplicitAnnotations(), a);
     }
 
     /**